requirements.txt
streamlit==1.34.0
selenium==4.16.0
beautifulsoup4==4.12.2
webdriver-manager==4.0.1
pandas==2.2.0
python-dotenv==1.0.0
matplotlib==3.8.2
numpy==1.26.2
scipy==1.11.4
openai==1.13.3
requests==2.31.0 